+++ src/components/gallery/GalleryView.tsx
@@ -1,0 +1,13 @@
+type GalleryViewProps = {
+  gallery: {
+    images: string[];
+    grid: number;
+    ratio: string;
+  };
+};
+const GalleryView = ({ gallery }: GalleryViewProps) => {
+  console.log(gallery);
+  return <div>GalleryView</div>;
+};
+
+export default GalleryView;
